refactor(native): consolidate platform-specific service logic
Merged platform-specific service logic into unified modules. Removed `macos_service_main.rs` and `windows_service_main.rs`, replacing them with `service_main.rs` that redirects to library functions. Updated the macOS and Windows-specific service implementations to reside within respective modules. Enhanced maintainability by centralizing logic.
